+++ client/src/components/Login.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import { Tabs, Row, Col } from 'antd';
 import Typist from 'react-typist';
-import TypistLoop from 'react-typist-loop';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 
@@ -29,17 +28,17 @@ class Login extends React.Component {
 				<Row>
 					<Col xs={0} sm={0} md={12}>
 						<div className="login-info">
-							<div>Find best tutorials from&nbsp;</div>
-							<TypistLoop>
-								{['Udemy', 'Medium', 'Udacity', 'Edx', 'Lynda', 'YouTube', 'Coursera'].map(
-									text => (
-										<Typist key={text} cursor={{ show: false }}>
-											<span>{text}</span>
-											<Typist.Backspace count={text.length} delay={1000} />
-										</Typist>
-									)
-								)}
-							</TypistLoop>
+							<Typist cursor={{ show: false }}>
+								<div className="login-header">Find the best tutorials from</div>
+								<div className="websites">Udemy</div>
+								<div className="websites">Udacity</div>
+								<div className="websites">Edx</div>
+								<div className="websites">Lynda</div>
+								<div className="websites">Medium</div>
+								<div className="websites">YouTube</div>
+								<div className="websites">Coursera</div>
+								<div className="websites">and many more...</div>
+							</Typist>
 						</div>
 					</Col>
 					<Col xs={24} sm={24} md={12}>