+++ client/src/components/TutorialCard.js
@@ -1,10 +1,11 @@
 import React from 'react';
-import { Card, Tag, Skeleton, Tooltip, Icon, Row, Col, Button, message, Popconfirm } from 'antd';
+import { Card, Tag, Skeleton, Tooltip, Icon, Row, Col, Button, message, Popconfirm, Badge } from 'antd';
 import { withRouter } from 'react-router-dom';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 
 import { addToFavorites, removeFromFavorites, clearMessage } from '../actions/userActions';
+import { addUpvote, removeUpvote, clearUpvoteMessage } from '../actions/tutorialActions';
 import { getUserProfile } from '../actions/authActions';
 
 import '../styles/TutorialCard.css';
@@ -13,13 +14,10 @@ class TutorialCard extends React.Component {
 	constructor(props) {
 		super(props);
 
-		this.state = {
-			upvote: false
-		};
-
-		this.changeUpvoteColor = this.changeUpvoteColor.bind(this);
 		this.addToFavorites = this.addToFavorites.bind(this);
 		this.removeFromFavorites = this.removeFromFavorites.bind(this);
+		this.addUpvote = this.addUpvote.bind(this);
+		this.removeUpvote = this.removeUpvote.bind(this);
 	}
 
 	addToFavorites() {
@@ -40,9 +38,22 @@ class TutorialCard extends React.Component {
 		setTimeout(() => this.props.clearMessage(), 3000);
 	}
 
-	changeUpvoteColor() {
-		const upvote = this.state.upvote;
-		this.setState({ upvote: !upvote });
+	addUpvote() {
+		if (!this.props.auth.authenticated) {
+			return message.info('You need to login to add to favorites');
+		}
+
+		this.props.addUpvote(this.props.tutorial._id);
+		message.success('Upvote Added');
+		this.props.getUserProfile();
+		setTimeout(() => this.props.clearUpvoteMessage(), 3000);
+	}
+
+	removeUpvote() {
+		this.props.removeUpvote(this.props.tutorial._id);
+		message.success('Upvote Removed');
+		this.props.getUserProfile();
+		setTimeout(() => this.props.clearUpvoteMessage(), 3000);
 	}
 
 	render() {
@@ -71,9 +82,29 @@ class TutorialCard extends React.Component {
 		}
 
 		let favorite = false;
+		let upvote = false;
+		let upvoteCount = 0;
+		if (this.props.tutorial.upvotes) upvoteCount = this.props.tutorial.upvotes.length;
+
 		if (this.props.auth.userProfile.favorites) {
 			this.props.auth.userProfile.favorites.forEach(tutorial => {
-				if (this.props.tutorial._id === tutorial._id) favorite = true;
+				if (this.props.tutorial._id === tutorial._id) {
+					favorite = true;
+					return;
+				} else {
+					favorite = false;
+				}
+			});
+		}
+
+		if (this.props.auth.userProfile.upvotes) {
+			this.props.auth.userProfile.upvotes.forEach(tutorial => {
+				if (this.props.tutorial._id === tutorial) {
+					upvote = true;
+					return;
+				} else {
+					upvote = false;
+				}
 			});
 		}
 
@@ -93,35 +124,61 @@ class TutorialCard extends React.Component {
 						</Tooltip>
 					</div>
 					<Row gutter={{ sm: 0, md: 4, xl: 8 }}>
-						<Col sm={{ span: 24 }} md={{ span: 12 }} xl={{ span: 8 }}>
+						<Col xs={24} sm={24} md={12} xl={8}>
 							<div className="card-entries">Medium : {this.props.tutorial.medium}</div>
 						</Col>
-						<Col sm={{ span: 24 }} md={{ span: 12 }} xl={{ span: 8 }}>
+						<Col xs={24} sm={24} md={12} xl={8}>
 							<div className="card-entries">Type : {this.props.tutorial.type}</div>
 						</Col>
-						<Col sm={{ span: 24 }} md={{ span: 12 }} xl={{ span: 8 }}>
+						<Col xs={24} sm={24} md={12} xl={8}>
 							<div className="card-entries">Skill Level : {this.props.tutorial.skillLevel}</div>
 						</Col>
 					</Row>
 					<div className="card-entries">{tags}</div>
-					{!favorite ? (
-						<Button type="primary" className="favorite-button" onClick={this.addToFavorites}>
-							Add to Favorites
-						</Button>
-					) : (
-						<Popconfirm
-							placement="top"
-							title="Remove from favorites?"
-							okText="Yes"
-							cancelText="Cancel"
-							icon={<Icon type="question-circle" theme="outlined" />}
-							onConfirm={this.removeFromFavorites}
-						>
-							<Button type="danger" className="remove-favorite-button">
-								Remove from Favorites
-							</Button>
-						</Popconfirm>
-					)}
+					<Row>
+						<Col xs={24} sm={24} md={18} lg={20} className="upvote-button">
+							{!upvote ? (
+								<Badge count={upvoteCount} showZero>
+									<Button className="upvote-button" onClick={this.addUpvote}>
+										Upvote
+									</Button>
+								</Badge>
+							) : (
+								<Popconfirm
+									placement="top"
+									title="Remove Upvote?"
+									okText="Yes"
+									cancelText="Cancel"
+									icon={<Icon type="question-circle" theme="outlined" />}
+									onConfirm={this.removeUpvote}
+								>
+									<Badge count={upvoteCount} showZero>
+										<Button type="danger" className="upvote-button">
+											Remove Upvote
+										</Button>
+									</Badge>
+								</Popconfirm>
+							)}
+						</Col>
+						<Col xs={24} sm={24} md={6} lg={4}>
+							{!favorite ? (
+								<Button type="primary" onClick={this.addToFavorites}>
+									Add to Favorites
+								</Button>
+							) : (
+								<Popconfirm
+									placement="top"
+									title="Remove from favorites?"
+									okText="Yes"
+									cancelText="Cancel"
+									icon={<Icon type="question-circle" theme="outlined" />}
+									onConfirm={this.removeFromFavorites}
+								>
+									<Button type="danger">Remove from Favorites</Button>
+								</Popconfirm>
+							)}
+						</Col>
+					</Row>
 				</Skeleton>
 			</Card>
 		);
@@ -142,5 +199,13 @@ const mapStateToProps = state => ({
 
 export default connect(
 	mapStateToProps,
-	{ addToFavorites, removeFromFavorites, clearMessage, getUserProfile }
+	{
+		addToFavorites,
+		removeFromFavorites,
+		clearMessage,
+		getUserProfile,
+		addUpvote,
+		removeUpvote,
+		clearUpvoteMessage
+	}
 )(withRouter(TutorialCard));