+++ client/src/reducers/tutorialReducer.js
@@ -1,10 +1,19 @@
-import { ADD_TUTORIAL, GET_TUTORIALS_BY_TAG, GET_TUTORIAL, TUTORIAL_LOADING } from '../actions/types';
+import {
+	ADD_TUTORIAL,
+	GET_TUTORIALS_BY_TAG,
+	GET_TUTORIAL,
+	TUTORIAL_LOADING,
+	ADD_UPVOTE,
+	REMOVE_UPVOTE,
+	CLEAR_UPVOTE_MESSAGE
+} from '../actions/types';
 
 const initialState = {
 	loading: false,
 	tutorial: null,
 	tutorials: [],
-	tag: ''
+	tag: '',
+	message: ''
 };
 
 export default function(state = initialState, action) {
@@ -30,6 +39,21 @@ export default function(state = initialState, action) {
 			return {
 				...state,
 				loading: true
+			};
+		case ADD_UPVOTE:
+			return {
+				...state,
+				message: action.payload.msg
+			};
+		case REMOVE_UPVOTE:
+			return {
+				...state,
+				message: action.payload.msg
+			};
+		case CLEAR_UPVOTE_MESSAGE:
+			return {
+				...state,
+				message: ''
 			};
 		default:
 			return state;