+++ config/keys_prod.js
@@ -1,4 +1,5 @@
 module.exports = {
 	mongoURI: process.env.MONGO_URI,
-	secretOrKey: process.env.SECRET_OR_KEY
+	secretOrKey: process.env.SECRET_OR_KEY,
+	emailVerificationKey: process.env.EMAIL_VERIFICATION_KEY
 };