+++ routes/api/tags.js
@@ -10,6 +10,7 @@ const Tag = require('../../models/Tag');
 // Desc		Returns list of all tags
 router.get('/', (req, res) => {
 	Tag.find({})
+		.sort({ tag: 1 })
 		.then(tags => {
 			res.json({ tags });
 		})