+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.1.0] - 2018-10-25
+
+### Added
+
+-   Upvotes Backend
+-   Alphabetically sorted Tutorials and Tags
+-   Email Verification
+
 ## [1.0.1] - 2018-10-17
 
 ### Bug Fixes